feature: add invitations to api

+++ backend/event-service/api/src/main/java/ru/calendorny/dto/ReminderDto.java
@@ -5,6 +5,6 @@ import java.util.List;
 
 public record ReminderDto(
     List<Integer> minutesBefore,
-    Integer reminderMethodId
+    Long reminderMethodId
 ) {
 }
